+++ src/module/synthesizer/EnvelopeADSR.cpp
@@ -8,6 +8,7 @@
 #include "EnvelopeADSR.h"
 #include "../../lib/os/file.h"
 #include "../../lib/math/math.h"
+#include <cmath>
 
 namespace tsunami {
 
@@ -57,7 +58,7 @@ void EnvelopeADSR::start(float volume) {
 	if (ttl_decay > 0)
 		step_decay = (value_sustain - value_peak) / (float)ttl_decay;
 	if (ttl_release > 0)
-		factor_release = exp( - 3.0f / (float)ttl_release);
+		factor_release = expf( - 3.0f / (float)ttl_release);
 
 
 	start_attack();
